fix: chat controller,settings service, blacklist controler

+++ src/main/java/com/titus/socialnetworkingsite2/services/ServiceImpl/BlackListServicesImpl.java
@@ -15,41 +15,47 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
-import static com.titus.socialnetworkingsite2.Dto.Response.ResponseConstants.SUCCESSFULLY_ADDED_TO_BLACKLIST;
-import static com.titus.socialnetworkingsite2.Dto.Response.ResponseConstants.SUCCESSFULLY_REMOVE_FROM_BLACKLIST;
+import static com.titus.socialnetworkingsite2.Dto.Response.ResponseConstants.*;
 
 @Service
 @RequiredArgsConstructor
 public class BlackListServicesImpl implements BlackListService {
 
     private final BlackListRepository blackListRepository;
+    private final UserRepository userRepository;
 
 
 
     @Override
     public GenResponse addToBlackList(BlackListDTO blackListDTO) {
 
+        // Current User
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         UserDetails user = (UserDetails) auth.getPrincipal();
 
 
-        // Check if user is already blacklisted using efficient query
-        if (blackListRepository.existsByBlacklistedAndBlacklistedBy(blackListDTO.getBlacklisted(), user.getUsername())) {
+        // check if user exist
+        var existingUser = userRepository.findByUsername(blackListDTO.getBlacklisted());
+        if (existingUser.isEmpty()) {
             return GenResponse.builder()
-                    .status(HttpStatus.OK.value())
-                    .message(blackListDTO.getBlacklisted() + " already blacklisted!")
+                    .status(HttpStatus.NOT_FOUND.value())
+                    .message(blackListDTO.getBlacklisted() + USER_NOT_FOUND)
+                    .build();
+        }
+        // Check if the recipient is blacklisted
+        Optional<BlackList> existingBlacklistedUser = blackListRepository.findByBlacklistedAndBlacklistedBy(blackListDTO.getBlacklisted(), user.getUsername());
+
+        if (existingBlacklistedUser.isPresent()) {
+            return GenResponse.builder()
+                    .status(HttpStatus.ACCEPTED.value())
+                    .message(ALREADY_BLACKLISTED + blackListDTO.getBlacklisted())
                     .build();
         }
 
-        System.out.println("blacklisted: " + blackListDTO.getBlacklisted());
-        System.out.println("blacklistedBy: " + blackListDTO.getBlacklistedBy());
-        System.out.println(blackListRepository.existsByBlacklistedAndBlacklistedBy(blackListDTO.getBlacklisted(), blackListDTO.getBlacklistedBy()));
 
-
-
-
-        // blacklist User
+        // Create and save the BlackList object
         BlackList blackList = BlackList.builder()
                 .blacklisted(blackListDTO.getBlacklisted())
                 .isBlacklisted(true)
@@ -57,20 +63,26 @@ public class BlackListServicesImpl implements BlackListService {
                 .blacklistedBy(user.getUsername())
                 .build();
         blackListRepository.save(blackList);
+
         return GenResponse.builder()
                 .status(HttpStatus.CREATED.value())
                 .message(SUCCESSFULLY_ADDED_TO_BLACKLIST).build();
     }
 
-
-
-
     // Remove Users from Blacklist
     @Override
-    public GenResponse removeFromBlacklist(BlackListDTO blackListDTO) {
+    public GenResponse removeFromBlacklist(String blacklisted) {
 
         // Retrieve the blacklisted user from the repository
-        BlackList blackListedUser = blackListRepository.findByBlacklisted(blackListDTO.getBlacklisted());
+        BlackList blackListedUser = blackListRepository.findByBlacklisted(blacklisted);
+
+        // Check if the user exists in the blacklist
+        if (blackListedUser == null) {
+            return GenResponse.builder()
+                    .status(HttpStatus.NOT_FOUND.value())
+                    .message(USER_NOT_FOUND_IN_BLACKLIST)
+                    .build();
+        }
 
         // Remove the user from the blacklist
         blackListRepository.delete(blackListedUser);
@@ -81,9 +93,7 @@ public class BlackListServicesImpl implements BlackListService {
     }
 
     @Override
-    public List<String> getBlacklists() {
-        return blackListRepository.findAllBlacklisted();
+    public List<BlackListDTO> getBlacklistsByUsername(String blacklistedBy) {
+        return blackListRepository.findAllByBlacklistedBy(blacklistedBy);
     }
-
-
 }
